+++ pypgx/api/pipeline.py
@@ -13,7 +13,7 @@ from . import utils, plot, genotype, core
 
 def run_chip_pipeline(
     gene, output, variants, panel=None, assembly='GRCh37', impute=False,
-    force=False
+    force=False, samples=None, exclude=False
 ):
     """
     Run PyPGx's genotyping pipeline for chip data.
@@ -24,14 +24,22 @@ def run_chip_pipeline(
         Target gene.
     output : str
         Output directory.
-    variants : str
-        VCF file (zipped or unzipped).
+    variants : str, optional
+        Input VCF file must be already BGZF compressed (.gz) and indexed
+        (.tbi) to allow random access. Statistical haplotype phasing will be
+        skipped if input VCF is already fully phased.
     assembly : {'GRCh37', 'GRCh38'}, default: 'GRCh37'
         Reference genome assembly.
     impute : bool, default: False
         If True, perform imputation of missing genotypes.
     force : bool, default : False
         Overwrite output directory if it already exists.
+    samples : str or list, optional
+        Subset the VCF for specified samples. This can be a text file (.txt,
+        .tsv, .csv, or .list) containing one sample per line. Alternatively,
+        you can provide a list of samples.
+    exclude : bool, default: False
+        If True, exclude specified samples.
     """
     if not core.is_target_gene(gene):
         raise core.NotTargetGeneError(gene)
@@ -42,14 +50,21 @@ def run_chip_pipeline(
     os.mkdir(output)
 
     imported_variants = utils.import_variants(gene, variants,
-        assembly=assembly, platform='Chip')
+        assembly=assembly, platform='Chip', samples=samples, exclude=exclude)
     imported_variants.to_file(f'{output}/imported-variants.zip')
-    phased_variants = utils.estimate_phase_beagle(imported_variants,
-        panel=panel, impute=impute)
-    phased_variants.to_file(f'{output}/phased-variants.zip')
-    consolidated_variants = utils.create_consolidated_vcf(
-        imported_variants, phased_variants)
-    consolidated_variants.to_file(f'{output}/consolidated-variants.zip')
+
+    # Skip statistical phasing if input VCF is already fully phased.
+    if imported_variants.type == 'VcfFrame[Consolidated]':
+        consolidated_variants = imported_variants
+    else:
+        phased_variants = utils.estimate_phase_beagle(
+            imported_variants, panel=panel)
+        phased_variants.to_file(f'{output}/phased-variants.zip')
+        consolidated_variants = utils.create_consolidated_vcf(
+            imported_variants, phased_variants)
+        consolidated_variants.to_file(
+            f'{output}/consolidated-variants.zip')
+
     alleles = utils.predict_alleles(consolidated_variants)
     alleles.to_file(f'{output}/alleles.zip')
     genotypes = genotype.call_genotypes(alleles=alleles)
@@ -64,8 +79,8 @@ def run_chip_pipeline(
 def run_ngs_pipeline(
     gene, output, variants=None, depth_of_coverage=None,
     control_statistics=None, platform='WGS', assembly='GRCh37', panel=None,
-    force=False, samples=None, do_not_plot_copy_number=False,
-    do_not_plot_allele_fraction=False
+    force=False, samples=None, exclude=False, samples_without_sv=None,
+    do_not_plot_copy_number=False, do_not_plot_allele_fraction=False
 ):
     """
     Run PyPGx's genotyping pipeline for NGS data.
@@ -73,7 +88,7 @@ def run_ngs_pipeline(
     During copy number analysis, if the input data is targeted sequencing,
     the method will apply inter-sample normalization using summary statistics
     across all samples. For best results, it is recommended to specify known
-    samples without SV using ``samples``.
+    samples without SV using ``samples_without_sv``.
 
     Parameters
     ----------
@@ -82,7 +97,9 @@ def run_ngs_pipeline(
     output : str
         Output directory.
     variants : str, optional
-        VCF file (zipped or unzipped).
+        Input VCF file must be already BGZF compressed (.gz) and indexed
+        (.tbi) to allow random access. Statistical haplotype phasing will be
+        skipped if input VCF is already fully phased.
     depth_of_coverage : str, optional
         Archive file or object with the semantic type
         CovFrame[DepthOfCoverage].
@@ -97,7 +114,13 @@ def run_ngs_pipeline(
         unzipped). By default, the 1KGP panel is used.
     force : bool, default : False
         Overwrite output directory if it already exists.
-    samples : list, optional
+    samples : str or list, optional
+        Subset the VCF for specified samples. This can be a text file (.txt,
+        .tsv, .csv, or .list) containing one sample per line. Alternatively,
+        you can provide a list of samples.
+    exclude : bool, default: False
+        If True, exclude specified samples.
+    samples_without_sv : list, optional
         List of known samples without SV.
     do_not_plot_copy_number : bool, default: False
         Do not plot copy number profile.
@@ -145,16 +168,25 @@ def run_ngs_pipeline(
 
     if small_var and variants is not None:
         imported_variants = utils.import_variants(gene, variants,
-            assembly=assembly, platform=platform)
+            assembly=assembly, platform=platform, samples=samples,
+            exclude=exclude)
         imported_variants.to_file(f'{output}/imported-variants.zip')
-        phased_variants = utils.estimate_phase_beagle(
-            imported_variants, panel=panel)
-        phased_variants.to_file(f'{output}/phased-variants.zip')
-        consolidated_variants = utils.create_consolidated_vcf(
-            imported_variants, phased_variants)
-        consolidated_variants.to_file(f'{output}/consolidated-variants.zip')
+
+        # Skip statistical phasing if input VCF is already fully phased.
+        if imported_variants.type == 'VcfFrame[Consolidated]':
+            consolidated_variants = imported_variants
+        else:
+            phased_variants = utils.estimate_phase_beagle(
+                imported_variants, panel=panel)
+            phased_variants.to_file(f'{output}/phased-variants.zip')
+            consolidated_variants = utils.create_consolidated_vcf(
+                imported_variants, phased_variants)
+            consolidated_variants.to_file(
+                f'{output}/consolidated-variants.zip')
+
         alleles = utils.predict_alleles(consolidated_variants)
         alleles.to_file(f'{output}/alleles.zip')
+
         if not do_not_plot_allele_fraction:
             os.mkdir(f'{output}/allele-fraction-profile')
             plot.plot_vcf_allele_fraction(
@@ -175,13 +207,19 @@ def run_ngs_pipeline(
         if isinstance(control_statistics, str):
             control_statistics = sdk.Archive.from_file(control_statistics)
 
+        if samples is not None:
+            control_statistics = utils.filter_samples(control_statistics,
+                samples=samples, exclude=exclude)
+
         control_statistics.check_type('SampleTable[Statistics]')
         control_statistics.check_metadata('Platform', platform)
         control_statistics.check_metadata('Assembly', assembly)
 
-        read_depth = utils.import_read_depth(gene, depth_of_coverage)
+        read_depth = utils.import_read_depth(gene, depth_of_coverage,
+            samples=samples, exclude=exclude)
         read_depth.to_file(f'{output}/read-depth.zip')
-        copy_number = utils.compute_copy_number(read_depth, control_statistics, samples=samples)
+        copy_number = utils.compute_copy_number(read_depth,
+            control_statistics, samples_without_sv=samples_without_sv)
         copy_number.to_file(f'{output}/copy-number.zip')
         cnv_calls = utils.predict_cnv(copy_number)
         cnv_calls.to_file(f'{output}/cnv-calls.zip')