+++ pypgx/cli/import_read_depth.py
@@ -31,9 +31,24 @@ def create_parser(subparsers):
         metavar='read-depth',
         help='Archive file with the semantic type CovFrame[ReadDepth].'
     )
+    parser.add_argument(
+        '--samples',
+        metavar='PATH',
+        nargs='+',
+        help='Specify which samples should be included for analysis \n'
+             'by providing a text file (.txt, .tsv, .csv, or .list) \n'
+             'containing one sample per line. Alternatively, you can \n'
+             'provide a list of samples.'
+    )
+    parser.add_argument(
+        '--exclude',
+        action='store_true',
+        help='Exclude specified samples.'
+    )
 
 def main(args):
     archive = utils.import_read_depth(
-        args.gene, args.depth_of_coverage
+        args.gene, args.depth_of_coverage, samples=args.samples,
+        exclude=args.exclude
     )
     archive.to_file(args.read_depth)