+++ pypgx/cli/plot_bam_copy_number.py
@@ -36,7 +36,10 @@ def create_parser(subparsers):
         '--samples',
         metavar='TEXT',
         nargs='+',
-        help='Create plots only for these samples.'
+        help='Specify which samples should be included for analysis \n'
+             'by providing a text file (.txt, .tsv, .csv, or .list) \n'
+             'containing one sample per line. Alternatively, you can \n'
+             'provide a list of samples.'
     )
     parser.add_argument(
         '--ymin',